+++ kcode/ALS_rolled.cl
@@ -78,7 +78,7 @@ static void inverseMatrix_CholeskyMethod(int n, __global float* A, __global floa
     }
 }
 
-static void Mt_byM_multiply_k(int i, int j, __global float* H, __global float* Result, const long ptr,
+static void Mt_byM_multiply_k(int i, int j, __global float* H, __global float* Result, const unsigned ptr,
                               __global const unsigned* idx) {
     int base = get_group_id(0) * j * j;
     int ss = get_local_id(0);
@@ -100,7 +100,7 @@ static void Mt_byM_multiply_k(int i, int j, __global float* H, __global float* R
 }
 
 __kernel void batchsolve(int i, int j, __global float* H, __global const float* val, __global float* result,
-                         __global const unsigned* colMajored_sparse_idx, __global const long* row_ptr,
+                         __global const unsigned* colMajored_sparse_idx, __global const unsigned* row_ptr,
                          __global const unsigned* col_idx) {
     int basev = get_group_id(0) * j;
     int ss = get_local_id(0);
@@ -189,7 +189,7 @@ __kernel void batchsolve(int i, int j, __global float* H, __global const float* 
 }
 
 __kernel void updateW_overH_kernel(const int rows,
-                                   __global const long* row_ptr,
+                                   __global const unsigned* row_ptr,
                                    __global const unsigned* col_idx,
                                    __global const unsigned* colMajored_sparse_idx,
                                    __global const float* val,
@@ -258,7 +258,7 @@ __kernel void updateW_overH_kernel(const int rows,
 }
 
 __kernel void updateH_overW_kernel(const int cols,
-                                   __global const long* col_ptr,
+                                   __global const unsigned* col_ptr,
                                    __global const unsigned* row_idx,
                                    __global const float* val,
                                    const float lambda,
